14-02-2024 Entrenamiento perceptron patrones Double en Java

+++ Java_eclipse/REDES_MONOCAPA/src/RED_NEURONAL_PERCEPTRON/PuertasLogicas.java
@@ -5,20 +5,20 @@ public class PuertasLogicas{
 	public static void main(String[] args) {
 		
 		// Puerta AND
-		int[][] entradas_AND = {{1,1,1},{1,1,0},{1,0,1},{1,0,0},{0,1,1},{0,1,0},{0,0,1},{0,0,0}};
-		int[][] salidas_AND = {{1},{0},{0},{0},{0},{0},{0},{0}};
+		double[][] entradas_AND = {{1,1,1},{1,1,0},{1,0,1},{1,0,0},{0,1,1},{0,1,0},{0,0,1},{0,0,0}};
+		double[][] salidas_AND = {{1},{0},{0},{0},{0},{0},{0},{0}};
 		
 		// Puerta AND
 		//int[][] entradas_AND = {{1,1},{1,0},{0,1},{0,0}};
 		//int[][] salidas_AND = {{1},{0},{0},{0}};
 		
 		// Puerta AND
-				int[][] entradas_OR = {{1,1,1},{1,1,0},{1,0,1},{1,0,0},{0,1,1},{0,1,0},{0,0,1},{0,0,0}};
-				int[][] salidas_OR = {{1},{1},{1},{1},{1},{1},{1},{0}};
+				double[][] entradas_OR = {{1,1,1},{1,1,0},{1,0,1},{1,0,0},{0,1,1},{0,1,0},{0,0,1},{0,0,0}};
+				double[][] salidas_OR = {{1},{1},{1},{1},{1},{1},{1},{0}};
 		
 		// Puerta NOR
-		int[][] entradas_NOT = {{0},{1}};
-		int[][] salidas_NOT = {{1},{0}};
+		double[][] entradas_NOT = {{0},{1}};
+		double[][] salidas_NOT = {{1},{0}};
 		
 		// Puerta OR
 		//int[][] entradas_OR = {{1,1},{1,0},{0,1},{0,0}};
@@ -30,7 +30,7 @@ public class PuertasLogicas{
 				
 	}
 	
-	public static void and(int[][] entradas, int[][] salidas) {
+	public static void and(double[][] entradas, double[][] salidas) {
 		
 		System.out.println("PUERTA AND...");
 		Perceptron perceptron = new Perceptron(entradas,salidas);
@@ -39,7 +39,7 @@ public class PuertasLogicas{
 
 	}
 	
-	public static void or(int[][] entradas, int[][] salidas) {
+	public static void or(double[][] entradas, double[][] salidas) {
 		
 		System.out.println("PUERTA OR...");
 		Perceptron perceptron = new Perceptron(entradas,salidas);
@@ -49,7 +49,7 @@ public class PuertasLogicas{
 		
 	}
 	
-	public static void not(int[][] entradas, int[][] salidas) {
+	public static void not(double[][] entradas, double[][] salidas) {
 		
 		System.out.println("PUERTA NOT...");
 		Perceptron perceptron = new Perceptron(entradas,salidas);
